Basic Calculation Program
This program has been completed, but we plan to optimize it further in the future. Upgrades will be introduced, so please wait until the program becomes more useful and effective. Kindly avoid attempting high-level tasks as this is only a prototype. The program is currently limited to performing simple calculations.